+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/Link/Extensions/Intent+Link.swift
@@ -22,6 +22,17 @@ extension Intent {
         return supportsLink(allowV2Features: false) && (linkFundingSources == [.bankAccount])
     }
 
+    var linkFlags: [String: Bool] {
+        switch self {
+        case .paymentIntent(let paymentIntent, _):
+            return paymentIntent.linkSettings?.linkFlags ?? [:]
+        case .setupIntent(let setupIntent, _):
+            return setupIntent.linkSettings?.linkFlags ?? [:]
+        case .deferredIntent(let elementsSession, _):
+            return elementsSession.linkSettings?.linkFlags ?? [:]
+        }
+    }
+
     var callToAction: ConfirmButton.CallToActionType {
         switch self {
         case .paymentIntent(_, let paymentIntent):