+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet.swift
@@ -152,6 +152,8 @@ public class PaymentSheet {
         ) { result in
             switch result {
             case .success(let intent, let savedPaymentMethods, let isLinkEnabled, let isApplePayEnabled):
+                let isCVCRecollectionEnabled = intent.cvcRecollectionEnabled
+
                 // Set the PaymentSheetViewController as the content of our bottom sheet
                 let presentPaymentSheetVC = {
                     let paymentSheetVC = PaymentSheetViewController(
@@ -160,6 +162,7 @@ public class PaymentSheet {
                         configuration: self.configuration,
                         isApplePayEnabled: isApplePayEnabled,
                         isLinkEnabled: isLinkEnabled,
+                        isCVCRecollectionEnabled: isCVCRecollectionEnabled,
                         delegate: self
                     )
 
@@ -167,6 +170,7 @@ public class PaymentSheet {
 
                     let updateBottomSheet: () -> Void = {
                         self.bottomSheetViewController.contentStack = [paymentSheetVC]
+                        paymentSheetVC.didFinishPresenting()
                     }
                     updateBottomSheet()
                 }
@@ -289,7 +293,10 @@ extension PaymentSheet: PaymentSheetViewControllerDelegate {
     func paymentSheetViewControllerFinishedOnPay(_ paymentSheetViewController: PaymentSheetViewController,
                                                  completion: (() -> Void)? = nil) {
         self.bottomSheetViewController.transitionSpinnerToComplete(animated: true) {
-            self.bottomSheetViewController.removeBlurEffect(animated: true, completion: completion)
+            self.bottomSheetViewController.removeBlurEffect(animated: true)
+            if let completion {
+                completion()
+            }
         }
     }
 
@@ -312,12 +319,10 @@ extension PaymentSheet: PaymentSheetViewControllerDelegate {
     func paymentSheetViewControllerDidSelectPayWithLink(
         _ paymentSheetViewController: PaymentSheetViewController
     ) {
-        self.bottomSheetViewController.addBlurEffect(animated: true) {
-            self.presentPayWithLinkController(
-                from: paymentSheetViewController,
-                intent: paymentSheetViewController.intent
-            )
-        }
+        self.presentPayWithLinkController(
+            from: paymentSheetViewController,
+            intent: paymentSheetViewController.intent
+        )
     }
 }
 
@@ -341,14 +346,16 @@ extension PaymentSheet: PayWithLinkWebControllerDelegate {
         intent: Intent,
         with paymentOption: PaymentOption
     ) {
-        self.bottomSheetViewController.startSpinner()
-        let psvc = self.findPaymentSheetViewController()
-        psvc?.clearTextFields()
-        psvc?.pay(with: paymentOption)
+        let backgroundColor = self.configuration.appearance.colors.background.withAlphaComponent(0.9)
+        self.bottomSheetViewController.addBlurEffect(animated: false, backgroundColor: backgroundColor) {
+            self.bottomSheetViewController.startSpinner()
+            let psvc = self.findPaymentSheetViewController()
+            psvc?.clearTextFields()
+            psvc?.pay(with: paymentOption)
+        }
     }
 
     func payWithLinkWebControllerDidCancel(_ payWithLinkWebController: PayWithLinkWebController) {
-        self.bottomSheetViewController.removeBlurEffect(animated: true)
     }
 
     private func findPaymentSheetViewController() -> PaymentSheetViewController? {
@@ -377,9 +384,7 @@ private extension PaymentSheet {
         )
 
         payWithLinkVC.payWithLinkDelegate = self
-        self.bottomSheetViewController.addBlurEffect(animated: true) {
-            payWithLinkVC.present(over: presentingController)
-        }
+        payWithLinkVC.present(over: presentingController)
     }
 
 }